requirements.txt
pyyaml==5.3
flask==1.1.2